+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 3.0.0
+* i : Code rewrite to avoid needing content script and clean unnecessary code
+* i : More optimized menu creation 
+* i : Use [Promise](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise) return version of [chrome.notifications.create](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/notifications/create)
+* i : Write clipboard using [navigator.clipboard.writeText](https://developer.mozilla.org/en-US/docs/Web/API/Clipboard/writeText)
+
 # 2.1.0
 * i : Update script and remove Dev version